Move the test code directory out of the src directory
