+++ package.json
@@ -11,13 +11,13 @@
     "url": "https://github.com/yhayase/skk-vscode"
   },
   "engines": {
-    "vscode": "^1.88.0"
+    "vscode": "^1.98.0"
   },
   "categories": [
     "Other"
   ],
   "activationEvents": [],
-  "main": "./out/extension.js",
+  "main": "./out/src/extension.js",
   "contributes": {
     "commands": [
       {
@@ -513,13 +513,15 @@
     }
   },
   "scripts": {
-    "vscode:prepublish": "npm run compile",
+    "vscode:prepublish": "npm run build",
     "compile": "tsc -p ./",
+    "build": "rm -rf ./out; tsc --project tsconfig.build.json",
     "watch": "tsc -watch -p ./",
     "pretest": "npm run compile && npm run lint",
     "lint": "eslint src --ext ts",
+    "test": "npm run test:unit; npm run test:integration",
     "test:unit": "npm run pretest && mocha",
-    "test:integration": "npm run pretest && node ./out/tests/integration/runTest.js"
+    "test:integration": "npm run pretest && node ./out/test/integration/runTest.js"
   },
   "devDependencies": {
     "@types/chai": "^4.3.16",
