+++ test/integration/suite/henkan/deletion-feature.test.ts
@@ -1,6 +1,6 @@
 import * as assert from 'assert';
 import * as vscode from 'vscode';
-import { getGlobalJisyo } from '../../../../lib/skk/jisyo/jisyo';
+import { getGlobalJisyo } from '../../../../src/lib/skk/jisyo/jisyo';
 import { expect } from 'chai';
 
 suite('候補削除機能において', async () => {
