+++ tsconfig.build.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./out/src",
+    "rootDir": "./src" 
+  },
+  "include": ["./src/**/*.ts"]
+}
